+++ src/instructions.h
@@ -172,7 +172,7 @@ class JumpIfFalseInstruction: public JumpInstruction {
 class CallInstruction: public Instruction {
     public:
     CallInstruction();
-    CallInstruction(const uint64_t& address, const uint64_t& param_count);
+    CallInstruction(const uint64_t& address, const uint8_t& param_count);
     void read(const std::vector<uint8_t>& buffer, uint64_t* index);
     void write(std::vector<uint8_t>& buffer) const;
     void execute(Vm& vm) const;
@@ -182,13 +182,13 @@ class CallInstruction: public Instruction {
 
     private:
     uint64_t address;
-    uint64_t param_count;
+    uint8_t param_count;
 };
 
 class RetInstruction: public Instruction {
     public:
     RetInstruction();
-    RetInstruction(const uint64_t& values_count);
+    RetInstruction(const uint8_t& values_count);
     void read(const std::vector<uint8_t>& buffer, uint64_t* index);
     void write(std::vector<uint8_t>& buffer) const;
     void execute(Vm& vm) const;
@@ -197,7 +197,7 @@ class RetInstruction: public Instruction {
     uint8_t size() const;
 
     private:
-    uint64_t values_count;
+    uint8_t values_count;
 };
 
 class DupInstruction: public Instruction {
